# [MCC-893601] Performance optimization for `accessible_operations`