@kjerinsky-mdsol 2022-03-25
Metrics observed from logs in test environment:

|_avg | _min | _max | _duration_ms_pct_50 | _duration_ms_pct_95 | _duration_ms_pct_99|
|-- | -- | -- | -- | -- | --|
|30.13954 | 6.03904 | 216.78592 | 22.29248 | 90.99008 | 127.16288|

Note: The metrics from background are based on round trip vs the value noted in the background are isolated to database execution only.